move default locale
