+++ src/contexts/locale/ILocaleContext.ts
@@ -48,42 +48,3 @@ export interface ICustomLocale {
     actions?: Partial<IActionLocale>;
     sider?: Partial<Record<string, string>>;
 }
-
-const capitalize = (str: string): string =>
-    str
-        .split(" ")
-        .map((s) => `${s[0].toUpperCase()}${s.slice(1)}`)
-        .join(" ");
-
-export const defaultLocale: ILocaleContext = {
-    common: {
-        ok: "OK",
-        cancel: "Cancel",
-        confirm: "Are you sure?",
-        delete: "Delete",
-        page: "Page",
-        nodata: "No Data",
-    },
-    auth: {
-        username: "Username",
-        password: "Password",
-        login: "Login",
-        logout: "Logout",
-    },
-    routes: {
-        list: (values) => capitalize(`List ${values?.resource}`),
-        edit: (values) => capitalize(`Edit ${values?.resource}`),
-        create: (values) => capitalize(`Create ${values?.resource}`),
-        view: (values) => capitalize(`View ${values?.resource}`),
-    },
-    actions: {
-        title: () => "Actions",
-        create: (values) => capitalize(`Create ${values?.resource}`),
-        edit: () => "Edit",
-        delete: () => "Delete",
-        save: () => "Save",
-    },
-    sider: {
-        dashboard: "Dashboard",
-    },
-};
